Hide lonely decimal point
The print_val() function uses the formatC() function with flag = "#".
Therefore, the decimal point is always displayed even if no fraction
digit follows it. Since this is unwanted formatC() is called without
flag = "#" if the number of significant figures is <= the number of
digits of the whole part of the (decimal) number.