Allow the assessment of single batches
Before, the expirest_osle() and expirest_wisle() functions worked only
with data sets having a batch_vbl factor with two or more levels. There
are situations, though, where the assessment should done with a single
batch. In the one-batch case, parts of the code can or must be skipped,
e.g. the code which checks for the most suitable model by aid of an
ANCOVA.
Currently, when the dids model applies, the results from fitting
individual models to each batch are reported. In principle, the dids
model is a model with individual intercepts and individual slopes
but pooled mean square error. This model was not reported so far.
With the current change, this model will be reported under the name
dids.pmse (for dids model with pooled mean square error). Furthermore,
the individual intercepts and points of intersection (of the confidence
interval limits with the corresponding specification limits) will be
reported, too.