Display ribbons going outside the canvas borders
Previously, when setting the ci_app option in plot_expirest_osle() or
plot_expirest_wisle() to 'ribbon', the corresponding ribbons may not be
shown in the parts of the plot where they go outside the canvas borders.
The use of the ggplot2 function coord_cartesian(), and the removal of
Specification of limits in the scaling functions, should prevent the
loss of information.