Allow selection of the model to be plotted
Currently, the model that was verified as the most appropriate model
(in terms of poolability) is plotted. Since it can be interesting to
see also the plots of the other models, an additional parameter must
be added to the parameter list which will allow to specify which model
should be printed if not the default model. On the other hand, the
parameter which allowed the user to choose whether the grouping should
be visible in the plot (i.e. different colours for the various batches)
becomes obsolete because when the option "no" was chosen to ignore the
grouping the basic model with common intercepts and common slopes was
shown.