Simplify use of x_/y_range
Currently, y_range must be provided. It will be made optional. If no value for y_range will be provided, the range will be determined based
on the data. Furthermore, the test if the lower and upper limit are
provided in the right order is omitted because it is easily possible
to determine which is the lower or the upper limit.