Perform the two-sided case on both sides
Currently, only the one-sided case is possible, i.e. in case of a
specification limit with two sides only one (the specified) side
is assessed. With the change, the two-sided case is performed
automatically. In addition, all tests are adapted to accept a
grouping factor (e.g. a batch variable) with only one factor level.