Correct bug in plot_expirest_wisle
The new functions get_text_annotation(), get_segments() and get_arrow()
stopped if the vector for the worst case intercepts (wc_icpt) was longer
than 1. However, the correct wc_icpt is selected from the vector by aid
of the parameter rl_index. Therefore, the test does not make sense and
was removed. In addition, further tests were added for the situation
where the wisle approach is performed with multiple rl values.