Reduce cyclomatic complexity
The estimation of predictions in plot_expirest_osle() and
plot_expirest_wisle() was moved to its own function. In addition, the
renaming of the ancillary data frames is now performed within the
corresponding construction functions. Since this required that even
more arguments were handed over, the interfaces of the corresponding
functions were reviewed and simplified.