Relax requirement on rl
Up to this change, the execution was sopped if ivl_side was lower but
any rl values was smaller than sl of if ivl_side was upper but any rl
value was greater than sl. In the change, the significant figures for
sl are taken into account when checking if an rl value is smaller or
greater than sl. Thus, e.g., if sl = 0.5 and sl_sf = 1, then rl = 0.54
is still accepted as because the rounded value (where sl_sf is the
number of significant figures) is not greater than 0.5. In addition,
checking of the relationship between rl and sl is only done in the
expirest_wisle() function but not any more in any of the helper
functions.