Correct usage of model_name being n.a.
In the plotting functions plot_expirest_wisle and plot_expirest_osle()
and the ancillary functions get_text_annotation(), get_vlines(),
get_segments() and get_arrow(), the model_name parameter is set to
dids if it is specified as n.a. In the generic summary functions, a
new variable is introduced so that dids is used as an index if the
model type acronym is n.a.