Let check_ancova convert cids decision to dids
The check_ancova() function was modified to take a decision in favour
of the different intercepts / different slopes (dids) if the ANCOVA
Concludes that the common intercepts / different slopes (cids) model
is appropriate because the latter model is not of practical relevance.
Text explaining the cids - dids topic was added to the relevant help
files.